Changes around Temporary  Surveys
A new field has been added to Survey to hold an indicator to identify if
the Survey is a Tempory one (Temporary_Survey). This appears on the
Licence tab of the Survey.  Previously temporary surveys were
identified by the Media Type, but at the time this was a bit of a fudge
to avoid adding a new field to Survey. The changes affect a number of
areas where temporary surveys are involved.