Import Wizard
Adds a new column to import wizard matching to show how the matches have
been generated and the number of matches found (Notes column) .  Allows
the users to select matches  at this stage without the need to go to the
hierarchy. Includes a number of changes to matching to prevent spurious
automatic matching.  This has been applied to Names/Species and
Locations but the code should allow  the 'Notes' column to be added to
any import field by adding new entries to the matching tables and
without further cahnges to core R^.