Mantis 715 - Identification  of inactive locations
This change was implemented in v6.29. using a udf to work out if a Location was inactive.  The use of the UDF is slowing down the population of Location hierarchy on large systems, This change adds an indicator to the Location table. This is still driven by Location Designation which is the way inactive locations have always been identified.