Mantis 716
In 6.29  the matching of names in the IW was tightened by not allowing potentially duplicate or invalid names into R6 when 'Make New Entries' was used. The SP used for this was also used by 'Make New Entrty' which prevented users creating an entry through the IW when they considered it was needed. This change provides seperate stored procedures. It has been applied to Name, but could also be applied to other columns, by making the appropriate entries in the IW Match tables.  '