Stops Population of Virtual Organism
The Taxon_List and Taxon_List Version for Virtual Organism will be remove at the next dictionary update opportunity. At the same time  the Seting table will be changed to update the entry for 'MasterList' to 'YES'. This will prevent the poulation of the Master List. At some point code around the management and use of the Virtual Lists can be removed but this shouldnt be done until most users are thought to be on Dictionary 0000004R or later.