Mantis 749
Stops need to use or populate Index_Taxon_Synonym. On system using the R6 Dictionary ITS was only being used to populate the details pane in Taxon Dictionary browser. The ITS table was reaching 7,000,000 records. Taxon Dictionary Browser details pane  is now populated from Index_Taxon_Name. ITS not populated. Note that this only happens if Setting 'SortMethod' is set to Organism. Old method is retained as an option  where the Dictionary does not have the Organism or NameServer tables.